Added Authentication to Journal Endpoints

+++ journalApp/src/main/java/com/learningSB/journalApp/service/JournalEntryService.java
@@ -20,36 +20,54 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
-    @Transactional  //this will make sure that ever line is executed like a single line of code as one and no line is left.
+    @Transactional
+    //this will make sure that ever line is executed like a single line of code as one and no line is left.
     public void saveEntry(JournalEntry journalEntry, String userName) {
         try {
-            User user= userService.findByUsername(userName);
+            User user = userService.findByUsername(userName);
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveNewUser(user);
         } catch (Exception e) {
             System.out.println(e);
-            throw new RuntimeException("Error saving journal entry",e);
+            throw new RuntimeException("Error saving journal entry", e);
         }
 
     }
+
     public void saveEntry(JournalEntry journalEntry) {
 
         journalEntryRepository.save(journalEntry);
     }
-public List<JournalEntry> getAll() {
-    return journalEntryRepository.findAll();
-}
 
-public Optional<JournalEntry> findById(ObjectId id) {
-    return journalEntryRepository.findById(id);
-}
+    public List<JournalEntry> getAll() {
+        return journalEntryRepository.findAll();
+    }
 
-public void deleteById(ObjectId id, String userName) {
-    User user= userService.findByUsername(userName);
-    user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-    userService.saveEntry(user);
-    journalEntryRepository.deleteById(id);
+    public Optional<JournalEntry> findById(ObjectId id) {
+        return journalEntryRepository.findById(id);
+    }
+
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUsername(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveNewUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("Error deleting journal entry", e);
+        }
+        return removed;
+    }
 }
-}
+/*public List<JournalEntry> findByUserName(String userName) {
+
+}*/
+
